recipes-devtools: Add recipe for PySensors module
The mechanism of reading the temperatures from the Thermal Monitoring
Unit was changed in the latest versions of the Linux Kernel. In order
to access the temperatures, the lm-sensors package has to be used, and
the telemetry scripts must be updated. To accomplish this from our
scripts, install the PySensors module and use it.

Issue: GVIP-1318
Signed-off-by: Bogdan Laiu <bogdan.laiu@nxp.com>
Signed-off-by: George Mocanu <george.mocanu@nxp.com>