layer: Drop meta-java usage
Drop the usage of the OpenJDK package provided via the meta-java
layer and use the Amazon Corretto distribution. In its current state,
the meta-java layer is no longer (actively) maintained, leading to
various issues in building the OpenJDK distribution (mainly around
fetching the dependencies).

Switch to using a prebuild distribution to speed the build times and
reduce the chance of build failures that might emerge during fetching
stage.

Issue: GVIP-1079
Signed-off-by: George Mocanu <george.mocanu@nxp.com>